+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Changelog
 
+##Version 0.8.4 (27/06/2013)
+* Add: Added `wake` option to send a WOL magic packet to a suspended/hibernating remote client. Specify the MAC address of the remote client in the `network.mac` property (ie. `network.mac=xx:xx:xx:xx:xx:xx`). When the client is no longer required, suspend or hibernate it with the `power` option.
+* Fix: Remove incorrect `qa`/`qax` dependency on Textures13.db
+
 ##Version 0.8.3 (17/06/2013)
 * Fix: Improve item matching performance on `watched` option